fix: connect mibs to postgres database
The mibs SQLAlchemy models now connect properly to the postgres
database.
init_db.sql is included in init_mibs_db.sh which creates a new database
for mibs that is separate from the db used by keycloak for auth.
The tables and database models  adhere to the schema specified in the
MIBS design doc.
Update mibs dependencies in dockerfile and requirements.txt for
SQLAlchemy.
Create mibs/db_test route that can test if database connection is setup
correctly.

closes #63