fix: add import for web project
the generated api can be imported by both reac-native and web.
the web project use craco due to issues that occur while importing the
local module.
switch the API generator to use typescript-axios instead of typescript
fetch. This is because of an error `Attempted import error: ... not exported
from './'` that was occuring while using the typescript-fetch generator.

closes #86