feat: implement authentication for flask services
The auth modlue now works as intended. Routes that are decorated with
`@auth.require_token` will now return a 401 response unless a
valid access token is provided with the request.

The auth module needs to be moved from `/src/projects/mibs/src` to
`/src/libs/flask-auth/` so that it is accessable to both mibs and cms.
This will likely require some configuration with docker and the python
path.

see: #69